refactor: move update notifier out of CLI layer
